+++ src/app/Playground/HåvardSamples/AnimalGroup.cs
@@ -1,0 +1,11 @@
+namespace Playground.HåvardSamples;
+
+public class AnimalGroup : List<Animal>
+{
+    public string Name { get; private set; }
+
+    public AnimalGroup(string name, List<Animal> animals) : base(animals)
+    {
+        Name = name;
+    }
+}